+++ web/app/api/hello/route.js
@@ -1,0 +1,5 @@
+import { NextResponse } from "next/server";
+
+export async function GET() {
+  return NextResponse.json({ name: "John Doe" });
+}
