+++ web/components/work-item-tile.js
@@ -1,3 +1,5 @@
+"use client";
+
 import Link from "next/link";
 import { useState } from "react";
 import urlForSanitySource from "../lib/urlForSanitySource";
