--- web/pages/_document.js
@@ -1,15 +1,0 @@
-import { Head, Html, Main, NextScript } from "next/document";
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head>
-        <link rel="stylesheet" href="https://use.typekit.net/gvj6hwn.css" />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
